move out versioner
