+++ orc8r/cloud/go/services/state/indexer/reindex/queue_sql.go
@@ -22,7 +22,6 @@ import (
 	"github.com/Masterminds/squirrel"
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
-	"github.com/thoas/go-funk"
 
 	"magma/orc8r/cloud/go/clock"
 	"magma/orc8r/cloud/go/services/state/indexer"
@@ -94,6 +93,7 @@ type sqlJobQueue struct {
 	maxAttempts uint
 	db          *sql.DB
 	builder     sqorc.StatementBuilder
+	IndexVersioner
 }
 
 // NewSQLJobQueue returns a new SQL-backed implementation of an unordered job queue.
@@ -116,7 +116,8 @@ type sqlJobQueue struct {
 //
 // Only provides Postgres support due to use of the non-standard "FOR UPDATE SKIP LOCKED" clause.
 func NewSQLJobQueue(maxAttempts uint, db *sql.DB, builder sqorc.StatementBuilder) JobQueue {
-	return &sqlJobQueue{maxAttempts: maxAttempts, db: db, builder: builder}
+	i :=  NewIndexVersioner(db, builder)
+	return &sqlJobQueue{maxAttempts: maxAttempts, db: db, builder: builder, IndexVersioner: i}
 }
 
 func (s *sqlJobQueue) Initialize() error {
@@ -209,7 +210,7 @@ func (s *sqlJobQueue) CompleteJob(job *Job, withErr error) error {
 
 		// Only update indexer actual versions on successful job completion
 		if withErr == nil {
-			err := s.setIndexerActualVersionImpl(tx, job.Idx.GetID(), job.To)
+			err := setIndexerActualVersionImpl(s.builder, tx, job.Idx.GetID(), job.To)
 			if err != nil {
 				return nil, err
 			}
@@ -261,26 +262,6 @@ func (s *sqlJobQueue) GetJobInfos() (map[string]JobInfo, error) {
 	}
 
 	return infos, nil
-}
-
-func (s *sqlJobQueue) GetIndexerVersions() ([]*indexer.Versions, error) {
-	txFn := func(tx *sql.Tx) (interface{}, error) {
-		return s.getIndexerVersionsImpl(tx)
-	}
-	txRet, err := sqorc.ExecInTx(s.db, &sql.TxOptions{Isolation: sql.LevelSerializable}, nil, txFn)
-	if err != nil {
-		return nil, err
-	}
-	ret := txRet.([]*indexer.Versions)
-	return ret, nil
-}
-
-func (s *sqlJobQueue) SetIndexerActualVersion(indexerID string, version indexer.Version) error {
-	txFn := func(tx *sql.Tx) (interface{}, error) {
-		return nil, s.setIndexerActualVersionImpl(tx, indexerID, version)
-	}
-	_, err := sqorc.ExecInTx(s.db, &sql.TxOptions{Isolation: sql.LevelSerializable}, nil, txFn)
-	return err
 }
 
 func (s *sqlJobQueue) initQueueTable() error {
@@ -345,7 +326,7 @@ func (s *sqlJobQueue) addJobs(tx *sql.Tx, newJobs []*reindexJob) error {
 // getNewJobs returns slice of reindex jobs to run.
 // Includes jobs where desired != actual.
 func (s *sqlJobQueue) getNewJobs(tx *sql.Tx) ([]*reindexJob, error) {
-	versions, err := s.getIndexerVersionsImpl(tx)
+	versions, err := getIndexerVersionsImpl(s.builder, tx)
 	if err != nil {
 		return nil, err
 	}
@@ -483,100 +464,6 @@ func (s *sqlJobQueue) selectAll() squirrel.SelectBuilder {
 	return s.builder.Select(idCol, fromCol, toCol, statusCol, attemptsCol, errorCol, lastChangeCol)
 }
 
-func (s *sqlJobQueue) getIndexerVersionsImpl(tx *sql.Tx) ([]*indexer.Versions, error) {
-	old, err := s.getTrackedVersions(tx)
-	if err != nil {
-		return nil, err
-	}
-
-	composed, err := getComposedVersions(old)
-	if err != nil {
-		return nil, err
-	}
-	if EqualVersions(composed, old) {
-		return composed, nil
-	}
-
-	// Test hook after first db call so the tx has "officially" started by acquiring some locks
-	TestHookGet()
-
-	err = s.overwriteAllVersions(tx, composed)
-	if err != nil {
-		return nil, err
-	}
-
-	return composed, nil
-}
-
-func (s *sqlJobQueue) setIndexerActualVersionImpl(tx *sql.Tx, indexerID string, version indexer.Version) error {
-	_, err := s.builder.Update(versionTableName).
-		Set(actualColVersions, version).
-		Where(squirrel.Eq{idColVersions: indexerID}).
-		RunWith(tx).
-		Exec()
-	if err != nil {
-		return errors.Wrapf(err, "update indexer actual version for %s to %d", indexerID, version)
-	}
-	return nil
-}
-
-func (s *sqlJobQueue) getTrackedVersions(tx *sql.Tx) ([]*indexer.Versions, error) {
-	var ret []*indexer.Versions
-
-	rows, err := s.builder.Select(idColVersions, actualColVersions, desiredColVersions).
-		From(versionTableName).
-		RunWith(tx).
-		Query()
-	if err != nil {
-		return nil, errors.Wrap(err, "get all indexer versions, select existing versions")
-	}
-
-	defer sqorc.CloseRowsLogOnError(rows, "GetAllIndexerVersions")
-
-	var idVal string
-	var actualVal, desiredVal int64 // int64 is driver's default int type, though these cols are actually int32 storing a uint32
-	for rows.Next() {
-		err = rows.Scan(&idVal, &actualVal, &desiredVal)
-		if err != nil {
-			return ret, errors.Wrap(err, "get all indexer versions, SQL row scan error")
-		}
-		v, err := newVersions(idVal, actualVal, desiredVal)
-		if err != nil {
-			return nil, err
-		}
-		ret = append(ret, v)
-	}
-
-	err = rows.Err()
-	if err != nil {
-		return ret, errors.Wrap(err, "get all indexer versions, SQL rows error")
-	}
-	sort.Slice(ret, func(i, j int) bool { return ret[i].IndexerID < ret[j].IndexerID }) // make deterministic
-	return ret, nil
-}
-
-func (s *sqlJobQueue) overwriteAllVersions(tx *sql.Tx, versions []*indexer.Versions) error {
-	_, err := s.builder.Delete(versionTableName).RunWith(tx).Exec()
-	if err != nil {
-		return errors.Wrap(err, "overwrite all indexer versions, delete existing versions")
-	}
-
-	if len(versions) == 0 {
-		return nil
-	}
-
-	builder := s.builder.Insert(versionTableName).Columns(idColVersions, actualColVersions, desiredColVersions)
-	for _, v := range versions {
-		builder = builder.Values(v.IndexerID, v.Actual, v.Desired)
-	}
-	_, err = builder.RunWith(tx).Exec()
-	if err != nil {
-		return errors.Wrapf(err, "overwrite all indexer desired versions, insert new versions %+v", versions)
-	}
-
-	return nil
-}
-
 // If no job available, returns ErrNotFound from magma/orc8r/lib/go/errors.
 func scanJob(rows *sql.Rows) (*reindexJob, error) {
 	jobs, err := scanJobs(rows)
@@ -618,69 +505,6 @@ func scanJobs(rows *sql.Rows) (map[string]*reindexJob, error) {
 	return jobs, nil
 }
 
-// getComposedVersions writes the composition of tracked (old) and local (new) indexers to store.
-// Determining whether an indexer needs to be reindexed depends on three recorded version infos per indexer:
-//	- new_desired	-- desired version from indexer registry
-//	- old_desired	-- desired version from existing reindex jobs
-//	- actual		-- actual version updated upon successful reindex job completion
-func getComposedVersions(old []*indexer.Versions) ([]*indexer.Versions, error) {
-	newv, err := getIndexerVersionsByID()
-	if err != nil {
-		return nil, err
-	}
-	composed := map[string]*indexer.Versions{}
-
-	// Insert all old versions -- old_desired and actual values
-	for _, v := range old {
-		composed[v.IndexerID] = v
-	}
-
-	// Insert all new versions -- new_desired overwrite any existing old_desired
-	for id, newDesired := range newv {
-		if _, present := composed[id]; present {
-			composed[id].Desired = newDesired
-		} else {
-			composed[id] = &indexer.Versions{IndexerID: id, Actual: 0, Desired: newDesired}
-		}
-	}
-
-	ret := funk.Map(composed, func(k string, v *indexer.Versions) *indexer.Versions { return v }).([]*indexer.Versions)
-	sort.Slice(ret, func(i, j int) bool { return ret[i].IndexerID < ret[j].IndexerID }) // make deterministic
-	return ret, nil
-}
-
-// getIndexerVersionsByID returns a map of registered indexer IDs to their registered ("desired") versions.
-func getIndexerVersionsByID() (map[string]indexer.Version, error) {
-	indexers, err := indexer.GetIndexers()
-	if err != nil {
-		return nil, err
-	}
-	ret := map[string]indexer.Version{}
-	for _, x := range indexers {
-		ret[x.GetID()] = x.GetVersion()
-	}
-	return ret, nil
-}
-
-// newVersions returns a new indexer versions view.
-// First checks the indexer versions fit in an indexer.Version.
-func newVersions(indexerID string, actualVersion, desiredVersion int64) (*indexer.Versions, error) {
-	actual, err := indexer.NewIndexerVersion(actualVersion)
-	if err != nil {
-		return nil, errors.Wrapf(err, "new actual version for indexer %s", indexerID)
-	}
-	desired, err := indexer.NewIndexerVersion(desiredVersion)
-	if err != nil {
-		return nil, errors.Wrapf(err, "new desired version for indexer %s", indexerID)
-	}
-	v := &indexer.Versions{
-		IndexerID: indexerID,
-		Actual:    actual,
-		Desired:   desired,
-	}
-	return v, nil
-}
-
 // reindexJob is the internal representation of a reindex job.
 type reindexJob struct {
 	// Indexer-relevant
@@ -696,10 +520,6 @@ type reindexJob struct {
 
 func (j *reindexJob) String() string {
 	return fmt.Sprintf("{id: %s, from: %d, to: %d}", j.id, j.from, j.to)
-}
-
-func (j *reindexJob) isSameVersions(job *reindexJob) bool {
-	return j.from == job.from && j.to == job.to
 }
 
 // getError for the reindex job.
